Add two tests that ensure the use of the default clock is time.Now() (#3)
Add two tests that ensure the use of the default clock is `time.Now()`.

In commit dfc1b68c I added an interface and made the now time used for
rate limiting something that defaulted to `time.Now()` but could be
injected for tests. There are significant tests for the `RateLimit`
function to ensure that it limits appropriately, but none tested that in
the absence of the injected test clock that `time.Now()` was still used.
These two tests add that coverage that I missed in the commit above.